tests: fix restAPI test fixture
